+++ src/utils/pasteUtils.js
@@ -2,11 +2,14 @@
 
 const removeSpecialChars = (event: SyntheticInputEvent<EventTarget>) => {
   const str = event.clipboardData.getData('Text');
-  const newStr = str.replace(/[^a-zA-Z0-9 `~!@#$%^&*()_|+\-=?;:'",.<>\{\}\[\]\\\/]\\n/gi, '');
+  const newStr = str.replace(
+    /[^a-zA-Z0-9 `~!@#$%^&*()_|+\-=?;:'",.<>\{\}\[\]\\\/]\\n/gi,
+    ''
+  );
   if (str !== newStr) {
-    console.log("Filtered unsupported characters!!");
+    console.log('Filtered unsupported characters!!');
     event.target.value = event.target.value + newStr;
-    event.preventDefault()
+    event.preventDefault();
   }
 };