+++ changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - IQVIACBEXU-632: Strange characters are displayed when content is copied and pasted from an external document 
   - IQVIACBEXU-557: Filter Proposals on All Tab does not filter if clicked away from
   - IQVIACBEXU-639: Proposal team history section giving blank value after delete the name from text field.
+  - IQVIACBEXU-788: Token refresh for timeout issues
   
 ### Changed
 ### Fixed